chyba skonczone

+++ app/src/main/java/com/example/bookmanager/data/model/Book.kt
@@ -4,7 +4,7 @@ data class Book(
     val id: String,
     val title: String,
     val author: String,
-    val coverId: Int?,
+    val coverId: Long?,
     val publishYear: Int?,
     val publishDate: String?,
     val description: String? = null
